New mode: "mixed" Legendre supporting jackknives

+++ python/cat_subsets_of_integrals.py
@@ -20,7 +20,10 @@ ns_samples = [int(s) for s in sys.argv[4:-1:2]]
 output_root = str(sys.argv[-1])
 collapse_factor = int(input_roots.pop()) if len(input_roots) > len(ns_samples) else 1 # recover the collapse factor if present
 assert collapse_factor > 0, "Collapsing factor must be positive"
+assert len(input_roots) >= 1, "Need at least one input directory"
 assert len(ns_samples) == len(input_roots), "Number of input dirs and subsamples to use from them must be the same"
+assert collapse_factor == 1 or not os.path.samefile(input_roots[0], output_root), "Only can collapse samples into a different directory"
+assert not any(os.path.samefile(input_root, output_root) for input_root in input_roots[1:]), "Only first input directory can be the same as the output"
 
 n_samples_tot = sum(ns_samples)
 assert n_samples_tot % collapse_factor == 0, "Collapse factor must divide the total number of samples"
@@ -54,8 +57,9 @@ for ii in range(len(I1)): # loop over all field combinations
     # read
     for input_root_all, n_samples in zip(input_roots_all, ns_samples):
         read_all = True
-        if collapse_factor == 1 and not os.path.isfile(input_root_all+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples)):
-            # if don't want to collapse and there are no more samples than we are using, can read averages from full file
+        if collapse_factor == 1 and not os.path.isfile(input_root_all+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples)) and os.path.isfile(input_root_all+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples-1)):
+            # if don't want to collapse and there are no more and no less samples than we are using, can read averages from full file
+            # note: logic can still be broken by a gap in the samples, but there probably would not be a full file then
             try:
                 c2.append(np.loadtxt(input_root_all+'c2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
                 c3.append(np.loadtxt(input_root_all+'c3_n%d_%s_%s_full.txt' % (n, mstr, index3)))
@@ -91,6 +95,7 @@ for ii in range(len(I1)): # loop over all field combinations
     else: # can copy files, which should be faster
         c2, c3, c4 = [None] * 3 # won't be needed, so can free memory
         for input_root_all, n_samples, sample_offset in zip(input_roots_all, ns_samples, sample_offsets):
+            if os.path.samefile(input_root_all, output_root_all): continue # skip copying, files are already there
             for i in trange(n_samples, desc="Copying %s full samples" % index4):
                 copy2(input_root_all+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_all+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
                 copy2(input_root_all+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i), output_root_all+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i + sample_offset))
@@ -99,21 +104,17 @@ for ii in range(len(I1)): # loop over all field combinations
 
     # jackknife integrals
     c2j, c3j, c4j = [], [], []
-    EEaA1, EEaA2, RRaA1, RRaA2 = [], [], [], []
     repeats = []
     # read
     for input_root_jack, n_samples in zip(input_roots_jack, ns_samples):
         read_all = True
-        if collapse_factor == 1 and not os.path.isfile(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples)):
-            # if don't want to collapse and there are no more samples than we are using, can read averages from full file
+        if collapse_factor == 1 and not os.path.isfile(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples)) and os.path.isfile(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, n_samples-1)):
+            # if don't want to collapse and there are no more and no less samples than we are using, can read averages from full file
+            # note: logic can still be broken by a gap in the samples, but there probably would not be a full file then
             try:
                 c2j.append(np.loadtxt(input_root_jack+'c2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
                 c3j.append(np.loadtxt(input_root_jack+'c3_n%d_%s_%s_full.txt' % (n, mstr, index3)))
                 c4j.append(np.loadtxt(input_root_jack+'c4_n%d_%s_%s_full.txt' % (n, mstr, index4)))
-                EEaA1.append(np.loadtxt(input_root_jack+'EE1_n%d_%s_%s_full.txt' % (n, mstr, index2)))
-                EEaA2.append(np.loadtxt(input_root_jack+'EE2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
-                RRaA1.append(np.loadtxt(input_root_jack+'RR1_n%d_%s_%s_full.txt' % (n, mstr, index2)))
-                RRaA2.append(np.loadtxt(input_root_jack+'RR2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
                 repeats.append(n_samples)
                 read_all = False
             except (FileNotFoundError, IOError): pass
@@ -123,50 +124,87 @@ for ii in range(len(I1)): # loop over all field combinations
                     c2j.append(np.loadtxt(input_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
                     c3j.append(np.loadtxt(input_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i)))
                     c4j.append(np.loadtxt(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i)))
-                    # cxj components
-                    EEaA1.append(np.loadtxt(input_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
-                    EEaA2.append(np.loadtxt(input_root_jack+'EE2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
-                    RRaA1.append(np.loadtxt(input_root_jack+'RR1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
-                    RRaA2.append(np.loadtxt(input_root_jack+'RR2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
                     repeats.append(1)
             else: break # end loop if last c4 jack not found
     if len(c4j) == 0: continue # skip rest of the loop if no jack integral has been found
     if sum(repeats) < n_samples_tot:
         print("ERROR: some %s jack samples missing: expected %d, found %d" % (index4, n_samples_tot, sum(repeats)))
         continue # skip the rest of the loop
-    c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2, repeats = [np.array(a) for a in (c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2, repeats)]
+    c2j, c3j, c4j, repeats = [np.array(a) for a in (c2j, c3j, c4j, repeats)]
     # average and save
-    c2jf, c3jf, c4jf, EEaA1f, EEaA2f, RRaA1f, RRaA2f = [np.average(a, axis=0, weights=repeats) for a in (c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2)]
+    c2jf, c3jf, c4jf = [np.average(a, axis=0, weights=repeats) for a in (c2j, c3j, c4j)]
     np.savetxt(output_root_jack+'c2_n%d_%s_%s_full.txt' % (n, mstr, index2), c2jf)
     np.savetxt(output_root_jack+'c3_n%d_%s_%s_full.txt' % (n, mstr, index3), c3jf)
     np.savetxt(output_root_jack+'c4_n%d_%s_%s_full.txt' % (n, mstr, index4), c4jf)
-    # cxj components
-    np.savetxt(output_root_jack+'EE1_n%d_%s_%s_full.txt' % (n, mstr, index2), EEaA1f)
-    np.savetxt(output_root_jack+'EE2_n%d_%s_%s_full.txt' % (n, mstr, index2), EEaA2f)
-    np.savetxt(output_root_jack+'RR1_n%d_%s_%s_full.txt' % (n, mstr, index2), RRaA1f)
-    np.savetxt(output_root_jack+'RR2_n%d_%s_%s_full.txt' % (n, mstr, index2), RRaA2f)
     if collapse_factor > 1:
-        c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2 = [np.mean(a.reshape(n_samples_out, collapse_factor, *np.shape(a)[1:]), axis=1) for a in (c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2)] # average adjacent chunks of collapse_factor samples
+        c2j, c3j, c4j = [np.mean(a.reshape(n_samples_out, collapse_factor, *np.shape(a)[1:]), axis=1) for a in (c2j, c3j, c4j)] # average adjacent chunks of collapse_factor samples
         # write the collapsed data
         for i in trange(n_samples_out, desc="Writing %s jack samples" % index4):
             np.savetxt(output_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), c2j[i])
             np.savetxt(output_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i), c3j[i])
             np.savetxt(output_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i), c4j[i])
-            # cxj components
+    else: # can copy files, which should be faster
+        c2j, c3j, c4j = [None] * 3 # won't be needed, so can free memory
+        for input_root_jack, n_samples, sample_offset in zip(input_roots_jack, ns_samples, sample_offsets):
+            if os.path.samefile(input_root_jack, output_root_jack): continue # skip copying, files are already there
+            for i in trange(n_samples, desc="Copying %s jack samples" % index4):
+                copy2(input_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
+                copy2(input_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i), output_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i + sample_offset))
+                copy2(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i), output_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i + sample_offset))
+    print("Done with %s jack" % index4)
+
+    # disconnected jackknife integrals (cxj). Done separately because absent in mixed Jackknife (JACKKNIFE_MIX)
+    EEaA1, EEaA2, RRaA1, RRaA2 = [], [], [], []
+    repeats = []
+    # read
+    for input_root_jack, n_samples in zip(input_roots_jack, ns_samples):
+        read_all = True
+        if collapse_factor == 1 and not os.path.isfile(input_root_jack+'EE1_n%d_%s_%s_%d.txt' % (n, mstr, index2, n_samples)) and os.path.isfile(input_root_jack+'EE1_n%d_%s_%s_%d.txt' % (n, mstr, index2, n_samples-1)):
+            # if don't want to collapse and there are no more and no less samples than we are using, can read averages from full file
+            # note: logic can still be broken by a gap in the samples, but there probably would not be a full file then
+            try:
+                EEaA1.append(np.loadtxt(input_root_jack+'EE1_n%d_%s_%s_full.txt' % (n, mstr, index2)))
+                EEaA2.append(np.loadtxt(input_root_jack+'EE2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
+                RRaA1.append(np.loadtxt(input_root_jack+'RR1_n%d_%s_%s_full.txt' % (n, mstr, index2)))
+                RRaA2.append(np.loadtxt(input_root_jack+'RR2_n%d_%s_%s_full.txt' % (n, mstr, index2)))
+                repeats.append(n_samples)
+                read_all = False
+            except (FileNotFoundError, IOError): pass
+        if read_all:
+            if os.path.isfile(input_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, n_samples-1)):
+                for i in trange(n_samples, desc="Reading %s disconnected jack samples" % index2):
+                    EEaA1.append(np.loadtxt(input_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
+                    EEaA2.append(np.loadtxt(input_root_jack+'EE2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
+                    RRaA1.append(np.loadtxt(input_root_jack+'RR1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
+                    RRaA2.append(np.loadtxt(input_root_jack+'RR2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i)))
+                    repeats.append(1)
+            else: break # end loop if last c4 jack not found
+    if len(EEaA1) == 0: continue # skip rest of the loop if no jack integral has been found
+    if sum(repeats) < n_samples_tot:
+        print("ERROR: some %s disconnected jack samples missing: expected %d, found %d" % (index2, n_samples_tot, sum(repeats)))
+        continue # skip the rest of the loop
+    EEaA1, EEaA2, RRaA1, RRaA2, repeats = [np.array(a) for a in (EEaA1, EEaA2, RRaA1, RRaA2, repeats)]
+    # average and save
+    EEaA1f, EEaA2f, RRaA1f, RRaA2f = [np.average(a, axis=0, weights=repeats) for a in (EEaA1, EEaA2, RRaA1, RRaA2)]
+    np.savetxt(output_root_jack+'EE1_n%d_%s_%s_full.txt' % (n, mstr, index2), EEaA1f)
+    np.savetxt(output_root_jack+'EE2_n%d_%s_%s_full.txt' % (n, mstr, index2), EEaA2f)
+    np.savetxt(output_root_jack+'RR1_n%d_%s_%s_full.txt' % (n, mstr, index2), RRaA1f)
+    np.savetxt(output_root_jack+'RR2_n%d_%s_%s_full.txt' % (n, mstr, index2), RRaA2f)
+    if collapse_factor > 1:
+        EEaA1, EEaA2, RRaA1, RRaA2 = [np.mean(a.reshape(n_samples_out, collapse_factor, *np.shape(a)[1:]), axis=1) for a in (EEaA1, EEaA2, RRaA1, RRaA2)] # average adjacent chunks of collapse_factor samples
+        # write the collapsed data
+        for i in trange(n_samples_out, desc="Writing %s disconnected jack samples" % index2):
             np.savetxt(output_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), EEaA1[i])
             np.savetxt(output_root_jack+'EE2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), EEaA2[i])
             np.savetxt(output_root_jack+'RR1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), RRaA1[i])
             np.savetxt(output_root_jack+'RR2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), RRaA2[i])
     else: # can copy files, which should be faster
-        c2j, c3j, c4j, EEaA1, EEaA2, RRaA1, RRaA2 = [None] * 7 # won't be needed, so can free memory
+        EEaA1, EEaA2, RRaA1, RRaA2 = [None] * 4 # won't be needed, so can free memory
         for input_root_jack, n_samples, sample_offset in zip(input_roots_jack, ns_samples, sample_offsets):
-            for i in trange(n_samples, desc="Copying %s jack samples" % index4):
-                copy2(input_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'c2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
-                copy2(input_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i), output_root_jack+'c3_n%d_%s_%s_%s.txt' % (n, mstr, index3, i + sample_offset))
-                copy2(input_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i), output_root_jack+'c4_n%d_%s_%s_%s.txt' % (n, mstr, index4, i + sample_offset))
-                # cxj components
+            if os.path.samefile(input_root_jack, output_root_jack): continue # skip copying, files are already there
+            for i in trange(n_samples, desc="Copying %s disconnected jack samples" % index2):
                 copy2(input_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'EE1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
                 copy2(input_root_jack+'EE2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'EE2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
                 copy2(input_root_jack+'RR1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'RR1_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
                 copy2(input_root_jack+'RR2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i), output_root_jack+'RR2_n%d_%s_%s_%s.txt' % (n, mstr, index2, i + sample_offset))
-    print("Done with %s jack" % index4)
+    print("Done with %s disconnected jack" % index2)
